+++ src/konpy/routes.clj
@@ -32,13 +32,14 @@
       :post {:handler answer/answer!}}]]
    ["/admin" {:middleware [[m/wrap-admin]]}
     ["" {:get  {:handler admin/tasks}
-         :post {:handler admin/create!}}]
-    ["/new" {:get  {:handler admin/new}
-             :post {:handler admin/create!}}]
-    ["/edit/:e" {:get  admin/edit
-                 :post admin/edit!}]
-    ;; no delete?
-    ["/delete/:e" {:delete admin/delete!}]]
+         :post {:handler admin/upsert!}}]
+    ; ["/upsert" {;:get  {:handler admin/new}
+    ;             :post {:handler admin/upsert!}}]
+    ; ["/edit/:e" {:get  admin/edit
+    ;              :post admin/edit!}]
+    ; ;; no delete?
+    ;["/delete/:e" {:delete admin/delete!}]
+    ]
    ;
    ["/example"
     ["" {:get  {:handler example/example-page}